new and delete calls are nearly gone!
With the exception of a new that is executed on CommStatsQWidget because it is tighly coupled with the UI(for now),
all of the new and  delete calls have been replaced with c++14 smart pointers. From now on, new and delete
shall never be used unless it is ABSOLUTELY necessary, which I doubt will ever be the case. Raw pointers are fine, as long as one
does not call delete on them. Let the smart pointers do their thing!